Nest unit tests and rename `utility.py` -> `utils.py`
