+++ .devcontainer/devcontainer.json
@@ -13,6 +13,6 @@
             ]
         }
     },
-    "postStartCommand": "pip3 --disable-pip-version-check --no-cache-dir install -r requirements.txt --break-system-packages && python3 -m pytest tests/*",
+    "postStartCommand": "pip3 --disable-pip-version-check --no-cache-dir install -r requirements.txt --break-system-packages && python3 -m pytest tests/**/test_*.py",
     "remoteUser": "vscode"
 }
