+++ routes/picture.py
@@ -4,7 +4,7 @@ from flask import Blueprint, jsonify, request, send_from_directory
 
 from config import PICTURE_FOLDER
 from jwt_helper import token_required
-from utility import database_cursor
+from utils import database_cursor
 
 picture_blueprint = Blueprint("picture", __name__)
