+++ tests/utility/test_validate_password.py
@@ -1,4 +1,14 @@
-from utility import validate_password
+import pytest
+from argon2 import PasswordHasher
+
+from utils import validate_password
+
+ph = PasswordHasher()
+
+
+@pytest.fixture
+def sample_password():
+    return "SecurePass123!"
 
 
 def test_valid_passwords():
