+++ tests/utility/test_verify_password.py
@@ -2,7 +2,7 @@ import pytest
 from argon2 import PasswordHasher
 from argon2.exceptions import VerificationError, VerifyMismatchError
 
-from utility import hash_password, verify_password
+from utils import hash_password, verify_password
 
 ph = PasswordHasher()
 
@@ -10,13 +10,6 @@ ph = PasswordHasher()
 @pytest.fixture
 def sample_password():
     return "SecurePass123!"
-
-
-def test_hash_password(sample_password):
-    """Ensure hash_password returns a hashed password as a non-empty string"""
-    hashed_password = hash_password(sample_password)
-    assert isinstance(hashed_password, str)
-    assert len(hashed_password) > 0
 
 
 def test_verify_password(sample_password):
